Changes to be committed: 	modified:   css/main.css 	new file:   img/ExampleCards/ImageOne.png 	new file:   img/ExampleCards/ImageThree.png 	new file:   img/ExampleCards/ImageTwo.png 	modified:   index.html 	modified:   js/carCards.js 	new file:   js/carSearch.js 	modified:   js/main.js
A ton of changes since the last commit:

- Introduced changes to carousel function, so that now it behaves properly,
-- It now has a timeout function which prevents the button from performing its function
whilst the animation from previous iteration is still playing.
-- It now keeps track of the left and right directions instead of just the transition
which enables it to animate properly,
-- in the future I plan to create a function which will trigger the changes of slides
every few seconds to match original website effect.

-Introduced carSearch.js which contains a CarSearch class that contains the method
enabling the creation of necessary selectors and options for the websites search section.
-- it keeps track of which Make option in "Make selector" is selected to dynamically display the
options of Model selector.
-- it adds the element of the dom and wipes the previously inserted options from model selector
when necessary.

Also added some temporary images to img/ExampleCards folder for the sake of testing, later to be deleted.